+++ inc/plugins/qwizi/dvz_shoutbox_bot/Commands/Prune.php
@@ -11,15 +11,17 @@ class Qwizi_DVZSB_Commands_Prune implements Qwizi_DVZSB_Commands_Base
 
     public function doAction($text, $uid)
     {
-        if ($text == $this->bot->settings('commands_prefix').'prune') {
-            $this->bot->delete();
+        if ($this->bot->accessMod()) {
+            if ($text == $this->bot->settings('commands_prefix') . 'prune') {
+                $this->bot->delete();
+            }
 
-        }
-
-        if (preg_match('/^\\'. $this->bot->settings('commands_prefix') . preg_quote('prune') . '[\s]+(.*)$/', $text, $matches)) {
-            $user = $this->bot->getUserInfo($matches[1]);
-            $this->bot->delete("uid={$user['uid']}");
-            return $this->bot->shout("Usunięto wiadomości użytkownika {$user['username']}");
+            if (preg_match('/^\\' . $this->bot->settings('commands_prefix') . preg_quote('prune') . '[\s]+(.*)$/', $text, $matches)) {
+                $user = $this->bot->getUserInfoFromUid($uid);
+                $target = $this->bot->getUserInfoFromUsername($matches[1]);
+                $this->bot->delete("uid={$target['uid']}");
+                return $this->bot->shout("@\"{$user['username']}\" usunął wiadomości użytkownika @\"{$target['username']}\"");
+            }
         }
     }
 }
